pyproject.toml
[tool.poetry]
name = "david-api"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.8"
tortoise-orm = "^0.16.17"
asyncio = "^3.4.3"
pydantic = "^1.7.1"
fastapi = "^0.61.1"
uvicorn = "^0.12.2"
google-api-python-client = "^1.12.5"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
oauth2client = "^4.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
